Add makefile and package.json to document install
The installation and update to run this as a presentation was not really
implemented, so I have added the Makefile to assist in the documentation
of the commands required. I don't use npm so this is so I can remember.

Additionally in updating the package-lock.json I have removed some
security vulnerabilities from the dependencies, not that they are really
a problem in this application.